[#173954342] implementa confirmacion para eliminar tarifas

+++ src/app/pages/empresa/empresa.component.ts
@@ -133,7 +133,7 @@ export class EmpresaComponent implements OnInit {
   }
 
   eliminarTarifa(id:string){
-   // if(confirm("¿Desea eliminar la tarifa")){
+    if(confirm("¿Desea eliminar la tarifa")){
     //  console.log("simon");
       this.empresaServicio.deleteTarifa(id).subscribe(
         res=>{
@@ -141,7 +141,7 @@ export class EmpresaComponent implements OnInit {
         },
         err=>console.log(err)
       );
-   // }
+    }
   }
 
   editarFormTarifa(tarifa:any) {
